reinstates o.a.i.applib.fixturescripts.FixtureScript as superclass of v2's equiv
for backward compatibility with incode platform
